+++ packages/imperative/src/cmd/src/syntax/SyntaxValidator.ts
@@ -234,6 +234,8 @@ export class SyntaxValidator {
                     }
                     if (positional.type === "number") {
                         valid = this.validateNumeric(commandArguments[positional.name], positional, responseObject, true) && valid;
+                        // Convert to number for backwards compatability
+                        if (valid) { commandArguments[positional.name] = parseInt(commandArguments[positional.name]); }
                     }
 
                     if (!(positional.stringLengthRange == null) &&
@@ -373,11 +375,11 @@ export class SyntaxValidator {
                             commandArguments[optionDef.name]);
                     }
                 } else if (optionDef.type === "boolean") {
-                    valid = this.validateBoolean(commandArguments[optionDef.name], optionDef,
-                        responseObject) && valid;
+                    valid = this.validateBoolean(commandArguments[optionDef.name], optionDef, responseObject) && valid;
                 } else if (optionDef.type === "number") {
-                    valid = this.validateNumeric(commandArguments[optionDef.name], optionDef,
-                        responseObject) && valid;
+                    valid = this.validateNumeric(commandArguments[optionDef.name], optionDef, responseObject) && valid;
+                    // Convert to numbers for backwards compatibility
+                    if (valid) { commandArguments[optionDef.name] = parseInt(commandArguments[optionDef.name]); }
                 }
                 /**
                  * Validate that the option's value is valid json.